+++ src/front/js/pages/login.js
@@ -8,7 +8,8 @@ import { Login } from "../component/login";
 import { Register } from "../component/register";
 
 export const Login = () => {
-	const { store, actions } = useContext(Context);
+    const { store, actions } = useContext(Context);
+    
 
 	return (
 		<div>
@@ -17,31 +18,3 @@ export const Login = () => {
 	);
 };
 
-export const loginUser = () =>{
-    const [email, setEmail] = userState("");
-    const [password, setPassword] = userState("");
-
-    const handleSubmit = e => {
-        e.preventDefault();
-
-        const body = {
-            email:email,
-            password:password,
-        };
-
-        fetch("https://3001-indigo-angelfish-r697zt96.ws-us03.gitpod.io/login", {
-            method: "POST",
-            body: JSON.stringify(body),
-            headers:{
-                "Content-Type":"application/json"
-            }
-        })
-        .then(res = res.json())
-        .then(data => {
-            console.log(data);
-            setAuth(true);
-        })
-        .catch(err => console.log(err));
-    };
-    
-}
